🔬 test(LottoTest): 로또 번호 유효성 검사 및 구매 기능에 대한 단위 테스트 추가
